+++ integrations/rueidis/CHANGELOG.md
@@ -7,7 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
-## [v1.0.1] - 2026-06-15 - Glob-Escape the Cache Clear Prefix
+## [v1.0.1] - 2026-06-16 - Glob-Escape the Cache Clear Prefix
 
 ### Fixed